+++ lensless/recon/gd.py
@@ -1,5 +1,14 @@
+# #############################################################################
+# gradient_descent.py
+# =================
+# Authors :
+# Eric BEZZAM [ebezzam@gmail.com]
+# Julien SAHLI [julien.sahli@epfl.ch]
+# #############################################################################
+
+
 import numpy as np
-from lensless.recon import ReconstructionAlgorithm
+from lensless.recon.recon import ReconstructionAlgorithm
 import inspect
 
 try: