+++ scripts/on_device_capture.py
@@ -2,7 +2,7 @@
 Capture raw Bayer data or post-processed RGB data.
 
 ```
-python scripts/on_device_capture.py --legacy --exp 0.02 --rgb --sensor_mode 1
+python scripts/on_device_capture.py --legacy --exp 0.02 --sensor_mode 0
 ```
 
 See these code snippets for setting camera settings and post-processing
@@ -12,14 +12,17 @@ See these code snippets for setting camera settings and post-processing
 
 """
 
+import hydra
 import os
 import cv2
-import click
 import numpy as np
 from time import sleep
 from PIL import Image
-from lensless.util import bayer2rgb, get_distro, rgb2gray, resize
-from lensless.constants import RPI_HQ_CAMERA_CCM_MATRIX, RPI_HQ_CAMERA_BLACK_LEVEL
+from lensless.hardware.utils import get_distro
+from lensless.utils.image import bayer2rgb, rgb2gray, resize
+from lensless.hardware.constants import RPI_HQ_CAMERA_CCM_MATRIX, RPI_HQ_CAMERA_BLACK_LEVEL
+from fractions import Fraction
+import time
 
 
 SENSOR_MODES = [
@@ -36,58 +39,23 @@ SENSOR_MODES = [
 ]
 
 
-@click.command()
-@click.option(
-    "--fn",
-    default="test",
-    type=str,
-    help="File name for recorded image.",
-)
-@click.option(
-    "--exp",
-    default=0.5,
-    type=float,
-    help="Exposure time in seconds.",
-)
-@click.option(
-    "--config_pause",
-    default=2,
-    type=float,
-    help="Pause time for loading / setting camera configuration.",
-)
-@click.option(
-    "--sensor_mode",
-    default="0",
-    type=click.Choice(np.arange(len(SENSOR_MODES)).astype(str)),
-    help="{'off': 0, 'auto': 1, 'sunlight': 2, 'cloudy': 3, 'shade': 4, 'tungsten': 5, "
-    "'fluorescent': 6, 'incandescent': 7, 'flash': 8, 'horizon': 9}",
-)
-@click.option(
-    "--rgb", is_flag=True, help="Whether to reconstruct RGB data or return raw bayer data."
-)
-@click.option(
-    "--gray",
-    is_flag=True,
-    help="Get grayscale data from the Raspberry Pi.",
-)
-@click.option("--iso", default=100, type=int)
-@click.option(
-    "--sixteen",
-    is_flag=True,
-)
-@click.option(
-    "--nbits_out",
-    default=8,
-    type=int,
-    help="Number of bits for output. Only used if saving RGB data.",
-)
-@click.option(
-    "--legacy",
-    is_flag=True,
-    help="Whether to use legacy camera software, despite being on Bullseye.",
-)
-@click.option("--down", type=float, help="Factor by which to downsample output.", default=None)
-def capture(fn, exp, config_pause, sensor_mode, iso, sixteen, rgb, nbits_out, legacy, gray, down):
+@hydra.main(version_base=None, config_path="../configs", config_name="capture")
+def capture(config):
+
+    bayer = config.bayer
+    fn = config.fn
+    exp = config.exp
+    config_pause = config.config_pause
+    sensor_mode = config.sensor_mode
+    rgb = config.rgb
+    gray = config.gray
+    iso = config.iso
+    sixteen = config.sixteen
+    legacy = config.legacy
+    down = config.down
+    res = config.res
+    nbits_out = config.nbits_out
+
     # https://www.raspberrypi.com/documentation/accessories/camera.html#maximum-exposure-times
     # TODO : check which camera
     assert exp <= 230
@@ -123,8 +91,8 @@ def capture(fn, exp, config_pause, sensor_mode, iso, sixteen, rgb, nbits_out, le
             stdout=subprocess.PIPE,
             stderr=subprocess.PIPE,
         )
-        result = cmd.stdout.readlines()
-        error = cmd.stderr.readlines()
+        cmd.stdout.readlines()
+        cmd.stderr.readlines()
         os.remove(jpg_fn)
         os.system(f"exiftool {dng_fn}")
         print("\nJPG saved to : {}".format(jpg_fn))
@@ -132,73 +100,148 @@ def capture(fn, exp, config_pause, sensor_mode, iso, sixteen, rgb, nbits_out, le
     else:
         import picamerax.array
 
-        camera = picamerax.PiCamera(framerate=1 / exp, sensor_mode=sensor_mode)
-
-        # camera settings, as little processing as possible
-        camera.iso = iso
-        camera.shutter_speed = int(exp * 1e6)
-        camera.exposure_mode = "off"
-        camera.drc_strength = "off"
-        camera.image_denoise = False
-        camera.image_effect = "none"
-        camera.still_stats = False
-
-        sleep(config_pause)
-        awb_gains = camera.awb_gains
-        camera.awb_mode = "off"
-        camera.awb_gains = awb_gains
-
-        print("Resolution : {}".format(camera.MAX_RESOLUTION))
-        print("Shutter speed : {}".format(camera.shutter_speed))
-        print("ISO : {}".format(camera.iso))
-        print("Frame rate : {}".format(camera.framerate))
-        print("Sensor mode : {}".format(SENSOR_MODES[sensor_mode]))
-        print("Config load time : {} seconds".format(config_pause))
-        # keep this as it needs to be parsed from remote script!
-        red_gain = float(awb_gains[0])
-        blue_gain = float(awb_gains[1])
-        print("Red gain : {}".format(red_gain))
-        print("Blue gain : {}".format(blue_gain))
-
-        # capture data
-        stream = picamerax.array.PiBayerArray(camera)
-        camera.capture(stream, "jpeg", bayer=True)
         fn += ".png"
 
-        # get bayer data
-        if sixteen:
-            output = np.sum(stream.array, axis=2).astype(np.uint16)
-        else:
-            output = (np.sum(stream.array, axis=2) >> 2).astype(np.uint8)
+        if bayer:
 
-        if rgb or gray:
+            # if rgb:
+
+            #     camera = picamerax.PiCamera(framerate=1 / exp, sensor_mode=sensor_mode, resolution=res)
+            #     camera.iso = iso
+            #     # Wait for the automatic gain control to settle
+            #     sleep(config_pause)
+            #     # Now fix the values
+            #     camera.shutter_speed = camera.exposure_speed
+            #     camera.exposure_mode = "off"
+            #     g = camera.awb_gains
+            #     camera.awb_mode = "off"
+            #     camera.awb_gains = g
+
+            #     print("Resolution : {}".format(camera.resolution))
+            #     print("Shutter speed : {}".format(camera.shutter_speed))
+            #     print("ISO : {}".format(camera.iso))
+            #     print("Frame rate : {}".format(camera.framerate))
+            #     print("Sensor mode : {}".format(SENSOR_MODES[sensor_mode]))
+            #     # keep this as it needs to be parsed from remote script!
+            #     red_gain = float(g[0])
+            #     blue_gain = float(g[1])
+            #     print("Red gain : {}".format(red_gain))
+            #     print("Blue gain : {}".format(blue_gain))
+
+            #     # take picture
+            #     fn += ".png"
+            #     camera.capture(str(fn), bayer=False, resize=None)
+
+            # else:
+
+            camera = picamerax.PiCamera(framerate=1 / exp, sensor_mode=sensor_mode, resolution=res)
+
+            # camera settings, as little processing as possible
+            camera.iso = iso
+            camera.shutter_speed = int(exp * 1e6)
+            camera.exposure_mode = "off"
+            camera.drc_strength = "off"
+            camera.image_denoise = False
+            camera.image_effect = "none"
+            camera.still_stats = False
+
+            sleep(config_pause)
+            awb_gains = camera.awb_gains
+            camera.awb_mode = "off"
+            camera.awb_gains = awb_gains
+
+            print("Resolution : {}".format(camera.resolution))
+            print("Shutter speed : {}".format(camera.shutter_speed))
+            print("ISO : {}".format(camera.iso))
+            print("Frame rate : {}".format(camera.framerate))
+            print("Sensor mode : {}".format(SENSOR_MODES[sensor_mode]))
+            # keep this as it needs to be parsed from remote script!
+            red_gain = float(awb_gains[0])
+            blue_gain = float(awb_gains[1])
+            print("Red gain : {}".format(red_gain))
+            print("Blue gain : {}".format(blue_gain))
+
+            # capture data
+            stream = picamerax.array.PiBayerArray(camera)
+            camera.capture(stream, "jpeg", bayer=True)
+
+            # get bayer data
             if sixteen:
-                n_bits = 12  # assuming Raspberry Pi HQ
+                output = np.sum(stream.array, axis=2).astype(np.uint16)
             else:
-                n_bits = 8
-            output_rgb = bayer2rgb(
-                output,
-                nbits=n_bits,
-                blue_gain=blue_gain,
-                red_gain=red_gain,
-                black_level=RPI_HQ_CAMERA_BLACK_LEVEL,
-                ccm=RPI_HQ_CAMERA_CCM_MATRIX,
-                nbits_out=nbits_out,
-            )
+                output = (np.sum(stream.array, axis=2) >> 2).astype(np.uint8)
 
-            if down:
-                output_rgb = resize(output_rgb, 1 / down, interpolation=cv2.INTER_CUBIC)
+            if rgb or gray:
+                if sixteen:
+                    n_bits = 12  # assuming Raspberry Pi HQ
+                else:
+                    n_bits = 8
 
-            # need OpenCV to save 16-bit RGB image
-            if gray:
-                output_gray = rgb2gray(output_rgb)
-                output_gray = output_gray.astype(output.dtype)
-                cv2.imwrite(fn, output_gray)
+                if config.awb_gains is not None:
+                    red_gain = config.awb_gains[0]
+                    blue_gain = config.awb_gains[1]
+
+                output_rgb = bayer2rgb(
+                    output,
+                    nbits=n_bits,
+                    blue_gain=blue_gain,
+                    red_gain=red_gain,
+                    black_level=RPI_HQ_CAMERA_BLACK_LEVEL,
+                    ccm=RPI_HQ_CAMERA_CCM_MATRIX,
+                    nbits_out=nbits_out,
+                )
+
+                if down:
+                    output_rgb = resize(
+                        output_rgb[None, ...], 1 / down, interpolation=cv2.INTER_CUBIC
+                    )[0]
+
+                # need OpenCV to save 16-bit RGB image
+                if gray:
+                    output_gray = rgb2gray(output_rgb[None, ...])
+                    output_gray = output_gray.astype(output_rgb.dtype).squeeze()
+                    cv2.imwrite(fn, output_gray)
+                else:
+                    cv2.imwrite(fn, cv2.cvtColor(output_rgb, cv2.COLOR_RGB2BGR))
             else:
-                cv2.imwrite(fn, cv2.cvtColor(output_rgb, cv2.COLOR_RGB2BGR))
+                img = Image.fromarray(output)
+                img.save(fn)
+
         else:
-            img = Image.fromarray(output)
-            img.save(fn)
+
+            # returning non-bayer data
+
+            from picamerax import PiCamera
+
+            camera = PiCamera()
+            if res:
+                assert len(res) == 2
+            else:
+                res = np.array(camera.MAX_RESOLUTION)
+                if down is not None:
+                    res = (np.array(res) / down).astype(int)
+
+            # Wait for the automatic gain control to settle
+            time.sleep(config.config_pause)
+
+            if config.awb_gains is not None:
+                assert len(config.awb_gains) == 2
+                g = (Fraction(config.awb_gains[0]), Fraction(config.awb_gains[1]))
+                g = tuple(g)
+                camera.awb_mode = "off"
+                camera.awb_gains = g
+                time.sleep(0.1)
+
+            print("Capturing at resolution: ", res)
+            print("AWB gains: ", float(camera.awb_gains[0]), float(camera.awb_gains[1]))
+
+            try:
+                camera.resolution = tuple(res)
+                camera.capture(fn)
+            except ValueError:
+                raise ValueError(
+                    "Out of resources! Use bayer for higher resolution, or increase `gpu_mem` in `/boot/config.txt`."
+                )
 
         print("\nImage saved to : {}".format(fn))