+++ scripts/recon/admm.py
@@ -13,7 +13,7 @@ import time
 import pathlib as plib
 import matplotlib.pyplot as plt
 import numpy as np
-from lensless.io import load_data
+from lensless.utils.io import load_data
 from lensless import ADMM